+++ openrem/remapp/extractors/extract_common.py
@@ -28,6 +28,7 @@
 """
 
 import logging
+from django.db.models import ObjectDoesNotExist
 
 
 def get_study_check_dup(dataset, modality='DX'):
@@ -108,3 +109,132 @@ def get_study_check_dup(dataset, modality='DX'):
                 return 0
     # study exists, but event doesn't
     return this_study
+
+
+def ct_event_type_count(g):
+    """Count CT event types and record in GeneralStudyModuleAttr summary fields
+
+    :param g: GeneralStudyModuleAttr database table
+    :return: None - database is updated
+    """
+
+    g.number_of_axial = 0
+    g.number_of_spiral = 0
+    g.number_of_stationary = 0
+    g.number_of_const_angle = 0
+    try:
+        events = g.ctradiationdose_set.get().ctirradiationeventdata_set.order_by('pk')
+        g.number_of_axial += events.filter(ct_acquisition_type__code_value__exact='113804').count()
+        g.number_of_spiral += events.filter(ct_acquisition_type__code_value__exact='116152004').count()
+        g.number_of_spiral += events.filter(ct_acquisition_type__code_value__exact='P5-08001').count()
+        g.number_of_spiral += events.filter(ct_acquisition_type__code_value__exact='C0860888').count()
+        g.number_of_stationary += events.filter(ct_acquisition_type__code_value__exact='113806').count()
+        g.number_of_const_angle += events.filter(ct_acquisition_type__code_value__exact='113805').count()
+    except ObjectDoesNotExist:
+        return
+    g.save()
+
+
+def populate_mammo_agd_summary(g):
+    """Copy accumulated AGD to the GeneralStudyModuleAttr summary fields
+
+    :param g: GeneralStudyModuleAttr database table
+    :return: None - database is updated
+    """
+
+    logger = logging.getLogger('remapp.extractors')
+
+    try:
+        for breast in g.projectionxrayradiationdose_set.get().accumxraydose_set.get().\
+                accummammographyxraydose_set.order_by('pk'):
+            if breast.laterality.code_value in ['T-04020', '73056007', 'C0222600']:  # Right breast
+                g.total_agd_right = breast.accumulated_average_glandular_dose
+            elif breast.laterality.code_value in ['T-04030', '80248007', 'C0222601']:  # Left breast
+                g.total_agd_left = breast.accumulated_average_glandular_dose
+            elif breast.laterality.code_value in ['T-04080', '63762007', 'C0222605']:  # Left breast
+                g.total_agd_both = breast.accumulated_average_glandular_dose
+        g.save()
+    except (ObjectDoesNotExist, AttributeError):
+        g.total_agd_both = 0
+        g.save()
+        logger.warning(u"Study UID {0}. Unable to set summary total_agd values - total_agd_both set to 0.".format(g.study_instance_uid))
+
+
+def populate_dx_rf_summary(g):
+    """Copy accumulated DAP and RP dose for each plane into the GeneralStudyModuleAttr summary fields
+
+    :param g: GeneralStudyModuleAttr database table
+    :return: None - database is updated
+    """
+
+    logger = logging.getLogger('remapp.extractors')
+
+    try:
+        planes = g.projectionxrayradiationdose_set.get().accumxraydose_set.order_by('pk')
+        accum_int_a = planes[0].accumintegratedprojradiogdose_set.get()
+        g.total_dap_a = accum_int_a.dose_area_product_total
+        g.total_rp_dose_a = accum_int_a.dose_rp_total
+        g.total_dap_a_delta_weeks = accum_int_a.dose_area_product_total_over_delta_weeks
+        g.total_rp_dose_a_delta_weeks = accum_int_a.dose_rp_total_over_delta_weeks
+        if g.projectionxrayradiationdose_set.get().irradeventxraydata_set.all(
+                ).filter(acquisition_plane__code_value__exact="113620").count():
+            g.number_of_events_a = g.projectionxrayradiationdose_set.get().irradeventxraydata_set.all(
+                ).filter(acquisition_plane__code_value__exact="113620").count()
+        elif g.projectionxrayradiationdose_set.get().irradeventxraydata_set.all(
+                ).filter(acquisition_plane__code_value__exact="113622").count():
+            g.number_of_events_a = g.projectionxrayradiationdose_set.get().irradeventxraydata_set.all(
+                ).filter(acquisition_plane__code_value__exact="113622").count()
+        else:
+            g.number_of_events_a = g.projectionxrayradiationdose_set.get().irradeventxraydata_set.all(
+                ).filter(acquisition_plane__code_value__exact="113890").count()
+        try:
+            accum_int_b = planes[1].accumintegratedprojradiogdose_set.get()
+            g.total_dap_b = accum_int_b.dose_area_product_total
+            g.total_rp_dose_b = accum_int_b.dose_rp_total
+            g.total_dap_b_delta_weeks = accum_int_b.dose_area_product_total_over_delta_weeks
+            g.total_rp_dose_b_delta_weeks = accum_int_b.dose_rp_total_over_delta_weeks
+            g.number_of_planes = 2
+            g.number_of_events_b = g.projectionxrayradiationdose_set.get().irradeventxraydata_set.all(
+                ).filter(acquisition_plane__code_value__exact="113621").count()
+        except IndexError:
+            g.number_of_planes = 1
+            logger.debug(u"Study UID {0}. No second plane when setting summary DAP/RP values".format(
+                g.study_instance_uid))
+        g.save()
+        try:
+            g.total_dap = g.total_dap_a + g.total_dap_b
+        except TypeError:
+            if g.total_dap_a is not None:
+                g.total_dap = g.total_dap_a
+            elif g.total_dap_b is not None:
+                g.total_dap = g.total_dap_b
+            else:
+                g.total_dap = 0
+        g.save()
+    except (ObjectDoesNotExist, IndexError):
+        g.number_of_events_a = 0
+        g.save()
+        logger.warning(u"Study UID {0}. Unable to set summary total DAP and RP dose values. number_of_events_a set to 0.".format(
+            g.study_instance_uid))
+
+
+def populate_rf_delta_weeks_summary(g):
+    """Copy DAP and RP dose accumulated over delta weeks into the GeneralStudyModuleAttr summary fields
+
+    :param g: GeneralStudyModuleAttr database table
+    :return: None - database is updated
+    """
+
+    logger = logging.getLogger('remapp.extractors')
+
+    try:
+        # Both planes are added to the total, which is recorded in each plane. So only need to get the first one
+        g.total_dap_delta_weeks = g.projectionxrayradiationdose_set.get().accumxraydose_set.order_by('pk')[
+            0].accumintegratedprojradiogdose_set.get().dose_area_product_total_over_delta_weeks
+        g.total_rp_dose_delta_weeks = g.projectionxrayradiationdose_set.get().accumxraydose_set.order_by('pk')[
+            0].accumintegratedprojradiogdose_set.get().dose_rp_total_over_delta_weeks
+        g.save()
+    except (ObjectDoesNotExist, IndexError):
+        logger.warning(u"Study UID {0}. Unable to set summary delta weeks DAP and RP dose values".format(
+            g.study_instance_uid))
+