+++ openrem/remapp/extractors/mam.py
@@ -383,6 +383,7 @@ def _patientmoduleattributes(dataset, g):  # C.7.1.1
 
 def _generalstudymoduleattributes(dataset, g):
     from datetime import datetime
+    from remapp.extractors.extract_common import populate_mammo_agd_summary
     from remapp.models import PatientIDSettings
     from remapp.tools.get_values import get_value_kw, get_seq_code_meaning, get_seq_code_value, list_to_string
     from remapp.tools.dcmdatetime import get_date, get_time
@@ -417,6 +418,9 @@ def _generalstudymoduleattributes(dataset, g):
     _projectionxrayradiationdose(dataset, g)
     _patientstudymoduleattributes(dataset, g)
     _patientmoduleattributes(dataset, g)
+    populate_mammo_agd_summary(g)
+    g.number_of_events = g.projectionxrayradiationdose_set.get().irradeventxraydata_set.count()
+    g.save()
 
 
 def _test_if_mammo(dataset):
@@ -432,7 +436,7 @@ def _test_if_mammo(dataset):
 def _mammo2db(dataset):
     from time import sleep
     from random import random
-    from remapp.extractors.extract_common import get_study_check_dup
+    from remapp.extractors.extract_common import get_study_check_dup, populate_mammo_agd_summary
     from remapp.models import GeneralStudyModuleAttr
     from remapp.tools import check_uid
     from remapp.tools.get_values import get_value_kw
@@ -448,6 +452,10 @@ def _mammo2db(dataset):
         this_study = get_study_check_dup(dataset, modality='MG')
         if this_study:
             _irradiationeventxraydata(dataset, this_study.projectionxrayradiationdose_set.get())
+            populate_mammo_agd_summary(this_study)
+            this_study.number_of_events = this_study.projectionxrayradiationdose_set.get(
+                ).irradeventxraydata_set.count()
+            this_study.save()
 
     if not study_in_db:
         # study doesn't exist, start from scratch